Add basic JS div creation function 16x16 + flexbox
Basic JS script created where a 16x16 grid is created via a two nested
for loops. First loop creates a div and appends it to a fixed div on the
HTML with id= "outsideContainer". After the div is appended, a second
for loop creates a second set of divs and appends each to each outer div.

For now only innerContainer and pixel classes have been added for each
type of div respectively.

Note: when I started writing this commit message there was a bug where
resizing the webpage squeezed horizontally each div, but it doesn't
anymore.