Add pixel coloring and clear canvas button
When the mouse passes over a div "pixel", a class "colorPixel" is added
to it. So far, the mouse doesn't need to be pressed down, but this is
a feature that we hope to implement on future versions.
A clear canvas button removes the "colorPixel" class which resets the
background color to white.