Add scalability for grid changes
Changed the node list creation mode from querySelectorAll to
getElementsByName. The previews mode created a static node list and
prevented functionality when the grid changed size.
Now every time the grid change button is pressed events are added to
the new node list.

If the user tries to change the grid size to greater than 100 or less
than 1 a pop-up is shown and the grid size is set to 100 or 1,
respectively.