Add random colors function, change style from JS
Changed the method of color changing divs of adding a class to changing
the style directly from the script. This let us add the random colors
function. A random number is generated and multiplied by 16777215 (White
on decimal format) and then converted to hexadecimal with the toString
function. This random color function is called every time the mouse
passes over a div IF the button "Change to Random Colors" was previously
pressed setting a flag as true. By clicking the "Change to Black color"
the previously mentioned flag is set to false.

Clear canvas button behavior was also modified, from removing the
"coloPixel" class to changing the divs background color to white.